[#231] - Move AutocompleteEntityFilterConfigurator class
